anavi-light-controller-sw.ino: Append 5 chars to AP
Append the last 5 characters of the machine ID to the Access Point
(AP) name for initial configuration. This simplifies the initial
configuration process and prevents WiFi mismatches if you launch
several units of ANAVI Light Controller for the first time at the
same time. In similar cases with this feature (almost always) each
of them will create AP will different name. There is still a very
small chance that there might be units with the same last 5
characters of their machine ID.

Signed-off-by: Leon Anavi <leon@anavi.org>